requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
filelock==3.16.1
Flask==2.2.0
frozenlist==1.4.1
fsspec==2024.9.0
huggingface-hub==0.25.0
idna==3.10
imap-tools==0.5.2
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
multidict==6.1.0
mysql-connector-python==8.0.30
numpy==1.24.4
openai==0.28.0
packaging==24.1
pandas==1.4.3
protobuf==3.20.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
six==1.16.0
tokenizers==0.13.3
torch==1.13.0
tqdm==4.66.5
transformers==4.26.0
typing==3.7.4.3
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==2.2.2
yarl==1.12.1
zipp==3.20.2
